revert wkcocos save system

+++ Include/WkCocos/LocalData/Comp/LocalData.h
@@ -35,7 +35,7 @@ namespace WkCocos
 					{
 						//QUICK file reading ( similar to C code, but still using C++ streams )
 						ifs.seekg(0, std::ios::end);
-						size_t size = ifs.tellg();
+						size_t size = static_cast<size_t>(ifs.tellg());
 
 						m_contents.resize(size);
 						ifs.seekg(0);
